Adding HTML Tests
So this is adding a bunch of checks on the HTML we generate by adding a
custom test class to HTMLProofer. We test for the following:

- Is the course displaying the correct title?
- Does the course have a clear selected button?
- Is there a "planner div" on the page?
- Is there a div for each year on the course?
- Is there a credit counter for each year?
- Do all the modules in the course appear on the page? Only once?

- Do modules display the correct titles?
- Do modules display the correct code?
- Do modules display the correct credit value?
- Do modules link to the correct more info page?
- Do modules provide for the correct modules?
- Do modules require the correct modules?

That was all I could think of as I was going. I'm pretty sure there will
be a few things that I was missing, so if you spot something that isn't
being tested be sure to let me know :)